requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
faiss-cpu
sentence-transformers
pypdf
google-generativeai
python-dotenv
fastapi
uvicorn
pydantic